Renamed folder to sui-distributed-execution to better reflect scope + some more factoring
