+++ Cargo.toml
@@ -89,7 +89,7 @@ members = [
     "crates/sui-move-natives",
     "crates/sui-network",
     "crates/sui-node",
-    "crates/sui-simple-fullnode",
+    "crates/sui-distributed-execution",
     "crates/sui-open-rpc",
     "crates/sui-open-rpc-macros",
     "crates/sui-proc-macros",
