+++ crates/sui-distributed-execution/src/lib.rs
@@ -1,4 +1,12 @@
+use sui_adapter::adapter::MoveVM;
+use sui_adapter::{execution_engine, execution_mode};
+use sui_config::genesis::Genesis;
 use sui_config::NodeConfig;
+use sui_core::transaction_input_checker::get_gas_status_no_epoch_store_experimental;
+use sui_protocol_config::ProtocolConfig;
+use sui_types::base_types::ExecutionDigests;
+use sui_types::message_envelope::Message;
+use sui_types::messages::{InputObjectKind, InputObjects, TransactionDataAPI, VerifiedTransaction};
 use sui_types::storage::get_module_by_id;
 
 use anyhow::Result;
@@ -6,7 +14,7 @@ use move_binary_format::CompiledModule;
 use move_bytecode_utils::module_cache::GetModule;
 use move_core_types::{language_storage::ModuleId, resolver::ModuleResolver};
 use prometheus::Registry;
-use std::collections::HashMap;
+use std::collections::{HashMap, HashSet};
 use std::sync::Arc;
 use sui_core::authority::authority_per_epoch_store::AuthorityPerEpochStore;
 use sui_core::authority::epoch_start_configuration::EpochStartConfiguration;
@@ -20,6 +28,7 @@ use sui_core::storage::RocksDbStore;
 use sui_node::metrics;
 use sui_types::metrics::LimitsMetrics;
 use sui_types::sui_system_state::SuiSystemStateTrait;
+use sui_types::temporary_store::TemporaryStore;
 use sui_types::{
     base_types::{ObjectID, ObjectRef, VersionNumber},
     error::{SuiError, SuiResult},
@@ -47,8 +56,8 @@ pub struct SequenceWorkerState {
 
 impl SequenceWorkerState {
     pub async fn new(config: &NodeConfig) -> Self {
-        let registry_service = { metrics::start_prometheus_server(config.metrics_address) };
         let genesis = config.genesis().expect("Could not load genesis");
+        let registry_service = { metrics::start_prometheus_server(config.metrics_address) };
         let prometheus_registry = registry_service.default_registry();
         let metrics = Arc::new(LimitsMetrics::new(&prometheus_registry));
         let checkpoint_store = CheckpointStore::new(&config.db_path().join("checkpoints"));
@@ -159,6 +168,153 @@ impl SequenceWorkerState {
                     .expect("Could not get highest checkpoint");
             }
             println!("Done downloading");
+        }
+    }
+
+    pub fn get_watermarks(&self) -> (u64, u64) {
+        let highest_synced_seq = match self
+            .checkpoint_store
+            .get_highest_synced_checkpoint_seq_number()
+            .expect("error")
+        {
+            Some(highest) => highest,
+            None => 0,
+        };
+        let highest_executed_seq = match self
+            .checkpoint_store
+            .get_highest_executed_checkpoint_seq_number()
+            .expect("error")
+        {
+            Some(highest) => highest,
+            None => 0,
+        };
+        (highest_synced_seq, highest_executed_seq)
+    }
+}
+
+pub struct ExecutionWorkerState {
+    pub memory_store: MemoryBackedStore,
+    // protocol_config: &'a ProtocolConfig,
+    // move_vm: &'a Arc<MoveVM>,
+    // epoch_start_config: &'a Arc<EpochStartConfiguration>,
+    // reference_gas_price: u64,
+}
+
+impl ExecutionWorkerState {
+    pub fn new(// protocol_config: &'a ProtocolConfig,
+        // move_vm: &'a Arc<MoveVM>,
+        // epoch_start_config: &'a Arc<EpochStartConfiguration>,
+        // reference_gas_price: u64,
+    ) -> Self {
+        Self {
+            memory_store: MemoryBackedStore::new(),
+            // protocol_config,
+            // move_vm,
+            // epoch_start_config,
+            // reference_gas_price,
+        }
+    }
+
+    pub fn init_store(&mut self, genesis: &Genesis) {
+        for obj in genesis.objects() {
+            self.memory_store
+                .objects
+                .insert(obj.id(), (obj.compute_object_reference(), obj.clone()));
+        }
+    }
+
+    pub async fn execute_tx(
+        &mut self,
+        tx: &VerifiedTransaction,
+        tx_digest: &ExecutionDigests,
+        checkpoint_seq: u64,
+        protocol_config: &ProtocolConfig,
+        move_vm: &Arc<MoveVM>,
+        epoch_start_config: &Arc<EpochStartConfiguration>,
+        reference_gas_price: u64,
+        metrics: Arc<LimitsMetrics>,
+    ) {
+        let tx_data = tx.data().transaction_data();
+        let (kind, signer, gas) = tx_data.execution_parts();
+        let input_object_kinds = tx_data
+            .input_objects()
+            .expect("Cannot get input object kinds");
+
+        let mut input_object_data = Vec::new();
+        for kind in &input_object_kinds {
+            let obj = match kind {
+                InputObjectKind::MovePackage(id)
+                | InputObjectKind::SharedMoveObject { id, .. }
+                | InputObjectKind::ImmOrOwnedMoveObject((id, _, _)) => {
+                    self.memory_store.objects.get(&id).expect("Object missing?")
+                }
+            };
+            input_object_data.push(obj.1.clone());
+        }
+
+        let gas_status = get_gas_status_no_epoch_store_experimental(
+            &input_object_data,
+            tx_data.gas(),
+            protocol_config,
+            reference_gas_price,
+            &tx_data,
+        )
+        .await
+        .expect("Could not get gas");
+
+        let input_objects = InputObjects::new(
+            input_object_kinds
+                .into_iter()
+                .zip(input_object_data.into_iter())
+                .collect(),
+        );
+        let shared_object_refs = input_objects.filter_shared_objects();
+        let transaction_dependencies = input_objects.transaction_dependencies();
+
+        let temporary_store = TemporaryStore::new(
+            &self.memory_store,
+            input_objects,
+            *tx.digest(),
+            protocol_config,
+        );
+
+        let (inner_temp_store, effects, _execution_error) =
+            execution_engine::execute_transaction_to_effects::<execution_mode::Normal, _>(
+                shared_object_refs,
+                temporary_store,
+                kind,
+                signer,
+                &gas,
+                *tx.digest(),
+                transaction_dependencies,
+                move_vm,
+                gas_status,
+                &epoch_start_config.epoch_data(),
+                protocol_config,
+                metrics.clone(),
+                false,
+                &HashSet::new(),
+            );
+
+        // Critical check: are the effects the same?
+        if effects.digest() != tx_digest.effects {
+            println!("Effects mismatch at checkpoint {}", checkpoint_seq);
+            let old_effects = tx_digest.effects;
+            println!("Past effects: {:?}", old_effects);
+            println!("New effects: {:?}", effects);
+        }
+        assert!(
+            effects.digest() == tx_digest.effects,
+            "Effects digest mismatch"
+        );
+
+        // And now we mutate the store.
+        // First delete:
+        for obj_del in &inner_temp_store.deleted {
+            self.memory_store.objects.remove(obj_del.0);
+        }
+        for (obj_add_id, (oref, obj, _)) in inner_temp_store.written {
+            self.memory_store.objects.insert(obj_add_id, (oref, obj));
         }
     }
 }
